requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.9
django-appconf==1.0.2
django-celery==3.2.2
django-celery-email==2.0.0
django-ckeditor==5.4.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-extensions==2.0.6
django-filter==1.1.0
django-geoposition==0.3.1
django-geosimple==1.0.3
django-js-asset==1.0.0
django-mptt==0.9.0
django-rest-swagger==2.1.2
django-tastypie==0.14.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-recursive==0.1.2
fake-factory==9999.9.9
geopy==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pyfcm==1.4.5
python-dateutil==2.7.0
python-geohash==0.8.5
python-mimeparse==1.6.0
pytz==2017.3
requests==2.18.4
requests-toolbelt==0.8.0
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
